requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
djangorestframework==3.15.2
idna==3.7
pytest==8.4.2
pytest-django==4.11.1
requests==2.32.4
sqlparse==0.5.1
stripe==10.5.0
typing_extensions==4.12.2
urllib3==2.5.0
